requirements.txt.txt
fastapi
uvicorn
jinja2
sentence-transformers
faiss-cpu
numpy
requests
pymupdf
python-multipart
